ContinuationSuite

Documentation for ContinuationSuite.

ContinuationSuite.Broyden

ContinuationSuite.ContinuationParameters

ContinuationSuite.ContinuationProblem

ContinuationSuite.ContinuationSolution

ContinuationSuite.LUSolver

ContinuationSuite.Newton

ContinuationSuite.NonlinearProblem

ContinuationSuite.PseudoArcLength

ContinuationSuite.autodiff_jac

ContinuationSuite.continuation

ContinuationSuite.finite_diff_jac

ContinuationSuite.solve

ContinuationSuite.Broyden — Type

Broyden{T <: Real, LS <: LinearSolver} <: NonlinearEqSolver

The Broyden's method for solving nonlinear equations. The nonlinear problem is solved iteratively using Broyden's method, which is a quasi-Newton method.

Broyden's method approximates the Jacobian matrix and updates it at every iteration. The update formula for the Jacobian is

\[J_{n+1} = J_n + ((f_{n+1} - f_n) - J_n Δx_n) Δx_n' / (Δx_n' Δx_n)\]

where J_n is the Jacobian matrix at the nth iteration and Δx_n = x_{n+1} - x_n is the change in the solution.

Fields

ϵₓ::T = 1e-8: The tolerance for the norm of the successive iterations ||xₙ₊₁ - xₙ||/||xₙ||

ϵᵣ::T = 1e-8: The tolerance for the norm of the residual ||f(xₙ)||

maxiters::Int = 20: The maximum number of iterations.

lsolver::LS = LUSolver(): The linear solver to use to solve the linear system.

info::Vector{Float64} = [0.0, 0.0, 0.0]: The information about the nonlinear solver.

info[1]: The norm of the residual ||f(xₙ)||

info[2]: The norm of the successive iterations ||xₙ₊₁ - xₙ||/||xₙ||

info[3]: The number of iterations.

Examples

using NumericalContinuation

f(x, p) = [x[1]^2 + x[2]^2 - 1; x[1] + x[2] - 1]
jac(x, p) = [2x[1] 2x[2]; 1 1]

p = 0.0

prob = NonlinearProblem(f, p; jac = jac)

x0 = [0.25, 0.5]
sol = solve(prob, Broyden(), x0)
sol = solve(prob, Broyden(ϵₓ = 1e-10, maxiters = 100), x0) # with different parameters

ContinuationSuite.ContinuationParameters — Type

ContinuationParameters{T <: Real, P <: Predictor, NLS <: NonlinearEqSolver} <: AbstractCache

Continuation parameters for the continuation problem. The fields are

Fields

- `λmin::T`: minimum value of the continuation parameter `λ`. (default: -1.0)
- `λmax::T`: maximum value of the continuation parameter `λ`. (default: 1.0)
- `Δs::T`: continuation step size. (default: 0.1)
- `direction::Symbol`: direction of the continuation. Can be `:forward` or `:backward`. (default: `:forward`)
- `maxsteps::Int`: maximum number of continuation steps. (default: 100)
- `predictor::P`: predictor method. (default: `PseudoArcLength()`)
- `corrector::NLS`: corrector method. (default: `Newton()`)
- `verbose::Bool`: whether to print information about the continuation process or not. (default: `false`)
- `ncols::Int`: number of columns to use when printing information about the continuation process. (default: 2)

ContinuationSuite.ContinuationProblem — Type

ContinuationProblem(f, ContPars [, p]; jac = nothing, autodiff = true)

Defines a Continuation Problem

To define a continuation problem, you need to provide the function f and the continuation parameters ContPars. Optionally, you can provide the Jacobian jac and additional parameters in p. The autodiff keyword argument is used to determine whether to use automatic differentiation or not. The default is true. This setting is ignored if the jacobian is provided.

A continuation problem is of the form:

\[f(x, λ, p) = 0,\]

where p is a parameter vector, x is the solution vector and λ is the continuation parameter.

Fields

- `f`: The function defining the continuation problem
- `jac`: The Jacobian of the function defining the continuation problem
- `p`: The parameter vector
- `ContPars`: The continuation parameters
- `autodiff`: Whether to use automatic differentiation or not
- `f_org`: The original function provided by the user
- `jac_org`: The original Jacobian provided by the user

Example of usage

using NumericalContinuation

function f(x, λ, p)
    x₁, x₂ = x
    f₁ = x₁^2 + x₂^2 - λ
    f₂ = x₂^2 - 2x₁ + 1
    return [f₁, f₂]
end

function jac(x, λ, p)
    x₁, x₂ = x
    fₓ = [2x₁ 2x₂; -2 2x₂]
    fλ = [-1; 0]
    return [fₓ fλ]
end

cont_pars = ContinuationParameters(λmin = 0.0, λmax = 2.5, Δs = 0.25,
                                   direction = :backward, predictor = PseudoArcLength(),
                                   corrector = Newton(), verbose = false)

prob = ContinuationProblem(f, cont_pars; autodiff = false)
x₀, λ₀ = [1.0, -1.0], 2.0

sol = continuation(prob, x₀, λ₀)

ContinuationSuite.ContinuationSolution — Type

ContinuationSolution(u, prob::ContinuationProblem)

Constructs a solution object for a continuation problem.

The solution is represented by a vector of the parameter values λ and a matrix of solutions x with the different components.

Fields

x: The solution matrix.

λ: The parameter values.

prob: The continuation problem.

ncols: The number of columns of the solution matrix x printed in the terminal.

Example of usage

using NumericalContinuation

function f(x, λ, p)
    x₁, x₂ = x
    f₁ = x₁^2 + x₂^2 - λ
    f₂ = x₂^2 - 2x₁ + 1
    return [f₁, f₂]
end

cont_pars = ContinuationParameters(λmin = 0.0, λmax = 2.5, Δs = 0.25,
                                   direction = :backward, predictor = PseudoArcLength(),
                                   corrector = Newton(), verbose = false)

prob = ContinuationProblem(f, cont_pars; autodiff = false)
x₀, λ₀ = [1.0, -1.0], 2.0

sol = continuation(prob, x₀, λ₀)
x, λ = sol.x, sol.λ

ContinuationSuite.LUSolver — Type

LUSolver <: LinearSolver

The default linear solver just calls the backslash \ operator to solve the linear system

A * x = b
x = A \ b

ContinuationSuite.Newton — Type

Newton{T <: Real, LS <: LinearSolver} <: NonlinearEqSolver

The Newton Raphson method for solving nonlinear equations. The nonlinear problem is solved iteratively using

\[J(x_n) Δx_n = -f(x_n)\]

where J(x_n) is the Jacobian matrix of f(x_n) and Δx_n is the update to the solution at the nth iteration. The update is given by

\[x_{n+1} = x_n + Δx_n.\]

Fields

ϵₓ::T = 1e-8: The tolerance for the norm of the successive iterations ||xₙ₊₁ - xₙ||/||xₙ||

ϵᵣ::T = 1e-8: The tolerance for the norm of the residual ||f(xₙ)||

maxiters::Int = 20: The maximum number of iterations.

lsolver::LS = LUSolver(): The linear solver to use to solve the linear system.

info::Vector{Float64} = [0.0, 0.0, 0.0]: The information about the nonlinear solver.

info[1]: The norm of the residual ||f(xₙ)||

info[2]: The norm of the successive iterations ||xₙ₊₁ - xₙ||/||xₙ||

info[3]: The number of iterations.

Examples

using NumericalContinuation

f(x, p) = [x[1]^2 + x[2]^2 - 1; x[1] + x[2] - 1]
jac(x, p) = [2x[1] 2x[2]; 1 1]

p = 0.0

prob = NonlinearProblem(f, p; jac = jac)

x0 = [0.25, 0.5]
sol = solve(prob, Newton(), x0)
sol = solve(prob, Newton(ϵₓ = 1e-10, maxiters = 100), x0) # with different parameters

ContinuationSuite.NonlinearProblem — Type

NonlinearProblem(f, [, p]; jac = nothing, autodiff = true)

Defines a Nonlinear System of Equations Problem

To define a nonlinear problem, you need to provide the function f. Optionally, you can provide the Jacobian jac and additional parameters in p. The autodiff keyword argument is used to determine whether to use automatic differentiation or not. The default is true. This setting is ignored if the jacobian is provided.

A nonlinear problem is of the form:

\[f(x, p) = 0,\]

where p is a parameter vector and x is the solution vector.

Fields

- `f`: The function defining the nonlinear problem
- `jac`: The Jacobian of the function defining the nonlinear problem
- `p`: The parameter vector
- `autodiff`: Whether to use automatic differentiation or not
- `f_org`: The original function provided by the user
- `jac_org`: The original Jacobian provided by the user

Example of usage

using NumericalContinuation

f(x, p) = [x[1]^2 + x[2]^2 - 1; x[1] + x[2] - 1]
jac(x, p) = [2x[1] 2x[2]; 1 1]

p = 0.0

prob = NonlinearProblem(f)
prob = NonlinearProblem(f, p)
prob = NonlinearProblem(f, p; jac = jac)

ContinuationSuite.PseudoArcLength — Type

PseudoArcLength <: TangentPredictor

Uses pseudo-arclength continuation to make a prediction for the next solution.

From a known value of the solution uₛ₀, the prediction is updated by computing the tangent vector t and the next solution is computed as:

\[uₛ = uₛ₀ + t Δs,\]

where Δs is the continuation step size in the pseudo-arclength. The tangent vector t is obtained from:

\[t = \frac{(xₛ, λₛ)}{\|(xₛ, λₛ)\|}, \]

where xₛ, λₛ is the solution of the linear system:

\[fₓ xₛ + fλ λₛ = 0,\]

which is equivalent to finding the nullspace of the extended jacobian [fₓ, fλ].

The tangent vector of the previous step is used to determine the direction of the tangent vector t. If the dot product of the previous tangent vector and the current tangent vector is negative, the tangent vector is flipped.

ContinuationSuite.autodiff_jac — Method

autodiff_jac(f, x)

Compute the Jacobian of f at x using automatic differentiation.

ContinuationSuite.continuation — Method

continuation(prob::ContinuationProblem, x₀, λ₀)

Solve the continuation problem prob starting from the initial solution x₀ and the initial parameter value λ₀. Return a ContinuationSolution object containing the solution.

Example of usage

using NumericalContinuation

function f(x, λ, p)
    x₁, x₂ = x
    f₁ = x₁^2 + x₂^2 - λ
    f₂ = x₂^2 - 2x₁ + 1
    return [f₁, f₂]
end

cont_pars = ContinuationParameters(λmin = 0.0, λmax = 2.5, Δs = 0.25,
                                   direction = :backward, predictor = PseudoArcLength(),
                                   corrector = Newton(), verbose = false)

prob = ContinuationProblem(f, cont_pars; autodiff = false)
x₀, λ₀ = [1.0, -1.0], 2.0

sol = continuation(prob, x₀, λ₀)

ContinuationSuite.finite_diff_jac — Method

finite_diff_jac(f, x)

Compute the Jacobian of f at x using forward finite differences.

ContinuationSuite.solve — Method

solve(prob, method, x₀)

Solve the nonlinear problem prob using the nonlinear equation solver method, starting from the initial guess x₀. The nonlinear equation solver is used to solve the nonlinear equation f(x, p) = 0, where f is the residual function defined in prob.